pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "torch>=2.0", "ninja"]
build-backend = "setuptools.build_meta"

[project]
name = "rist"
version = "1.0.0"
description = "Learning SO(3)-Invariant Semantic Correspondence via Local Shape Transform"
authors = [
    {name = "Chunghyun Park"},
    {name = "Seungwook Kim"},
    {name = "Jaesik Park"},
    {name = "Minsu Cho"},
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0",
    "numpy",
    "scipy",
    "einops",
    "tqdm",
    "lightning>=2.0",
    "hydra-core>=1.3",
    "hydra-colorlog>=1.2",
    "rootutils",
    "rich",
    "wandb",
    "omegaconf",
    "gdown>=4.7.0",
    "matplotlib",
]

[project.urls]
Homepage = "https://github.com/chrockey/RIST"

[tool.setuptools.packages.find]
include = ["src*"]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu128"
explicit = true
name = "pytorch-cu128"

[tool.uv.sources]
torch = { index = "pytorch-cu128" }

[tool.black]
line-length = 99
